fix: merge public key distribution contracts
