+++ rpc/apis.go
@@ -134,25 +134,11 @@ func (r *RPC) getApi(apiModule string) rpc.API {
 			Service:   api.NewChainApi(r.ledger),
 			Public:    true,
 		}
-	case "verifier":
+	case "pkd":
 		return rpc.API{
-			Namespace: "verifier",
+			Namespace: "pkd",
 			Version:   "1.0",
-			Service:   api.NewVerifierApi(r.cfgFile, r.ledger),
-			Public:    true,
-		}
-	case "oracle":
-		return rpc.API{
-			Namespace: "oracle",
-			Version:   "1.0",
-			Service:   api.NewOracleApi(r.cfgFile, r.ledger),
-			Public:    true,
-		}
-	case "publisher":
-		return rpc.API{
-			Namespace: "publisher",
-			Version:   "1.0",
-			Service:   api.NewPublisherApi(r.cfgFile, r.ledger),
+			Service:   api.NewPublicKeyDistributionApi(r.cfgFile, r.ledger),
 			Public:    true,
 		}
 	default:
@@ -190,7 +176,6 @@ func (r *RPC) GetWSApis() []rpc.API {
 
 func (r *RPC) GetPublicApis() []rpc.API {
 	apiModules := []string{"ledger", "account", "net", "util", "wallet", "mintage", "contract", "sms", "pledge",
-		"rewards", "pov", "miner", "config", "debug", "destroy", "metrics", "rep", "chain", "verifier", "oracle",
-		"publisher"}
+		"rewards", "pov", "miner", "config", "debug", "destroy", "metrics", "rep", "chain", "pkd"}
 	return r.GetApis(apiModules...)
 }
